Clean and refactor
