+++ mutalyzer_retriever/retriever.py
@@ -2,6 +2,7 @@ import json
 from copy import deepcopy
 from functools import lru_cache
 from pathlib import Path
+
 import requests
 
 from . import parser
@@ -17,7 +18,7 @@ class NoReferenceError(Exception):
     def __init__(self, status, uncertain_sources):
         self.uncertain_sources = uncertain_sources
         message = ""
-        if uncertain_sources is not []:
+        if uncertain_sources != []:
             message = f"\n\nUncertain sources: {', '.join(uncertain_sources)}\n"
 
         for source in status.keys():
@@ -49,12 +50,14 @@ def _raise_error(status):
             and isinstance(status[source]["errors"][0], NameError)
         ):
             uncertain_sources.append(source)
-    if uncertain_sources is []:
+    if uncertain_sources == []:
         raise NoReferenceRetrieved
     raise NoReferenceError(status, uncertain_sources)
 
 
-def _fetch_unknown_source(reference_id, reference_type, reference_source, size_off=True, timeout=1):
+def _fetch_unknown_source(
+    reference_id, reference_type, reference_source, size_off=True, timeout=1
+):
 
     status = {"lrg": {"errors": []}, "ncbi": {"errors": []}, "ensembl": {"errors": []}}
 
@@ -68,9 +71,7 @@ def _fetch_unknown_source(reference_id, reference_type, reference_source, size_o
             return reference_content, "lrg", "lrg"
     else:
         status["lrg"]["errors"].append(
-            ValueError(
-                f"Lrg fetch does not support '{reference_type}' reference type."
-            )
+            ValueError(f"Lrg fetch does not support '{reference_type}' reference type.")
         )
 
     # NCBI
@@ -126,11 +127,11 @@ def retrieve_raw(
     elif reference_source == "ncbi":
         reference_content, reference_type = ncbi.fetch(
             reference_id, reference_type, timeout
-        )  
+        )
     elif reference_source in ["ensembl", "ensembl_tark", "ensembl_rest"]:
         reference_content, reference_type = ensembl.fetch(
-            reference_id,reference_type,reference_source,timeout
-            )
+            reference_id, reference_type, reference_source, timeout
+        )
     elif reference_source == "lrg":
         reference_content = lrg.fetch_lrg(reference_id, timeout=timeout)
         if reference_content:
@@ -160,14 +161,14 @@ def retrieve_model(
     reference_content, reference_type, reference_source = retrieve_raw(
         reference_id, reference_source, reference_type, size_off, timeout=timeout
     )
- 
+
     if reference_type == "lrg":
         model = parser.parse(reference_content, reference_type, reference_source)
         if model_type == "all":
             return model
-        elif model_type == "sequence":
+        if model_type == "sequence":
             return model["sequence"]
-        elif model_type == "annotations":
+        if model_type == "annotations":
             return model["annotations"]
     elif reference_type == "gff3":
         if model_type == "all":
@@ -192,7 +193,7 @@ def retrieve_model(
         return {
             "sequence": parser.parse(reference_content, "fasta"),
         }
-    
+
     elif reference_type == "json":
         if "ensembl" in reference_source:
             json_model = parser.parse(reference_content, "json")
@@ -202,6 +203,7 @@ def retrieve_model(
                 return json_model["annotations"]
             elif model_type == "sequence":
                 return json_model["sequence"]["seq"]
+
 
 def retrieve_model_from_file(paths=[], is_lrg=False):
     """
