+++ tests/test_retriever_model.py
@@ -2,9 +2,10 @@ import json
 from pathlib import Path
 
 import pytest
-from mutalyzer_retriever import retrieve_model
-from .commons import references, _retrieve_raw, patch_retriever, _get_content
 
+from mutalyzer_retriever import retrieve_model
+
+from .commons import _get_content, patch_retriever, references
 
 
 def get_tests(references):
@@ -41,7 +42,9 @@ def _seq_from_rest(r_id):
     return _get_content("data/" + str(r_id) + ".sequence")
 
 
-@pytest.mark.parametrize("r_id, r_source, r_type, expected_model", get_tests(references))
+@pytest.mark.parametrize(
+    "r_id, r_source, r_type, expected_model", get_tests(references)
+)
 def test_model(r_id, r_source, r_type, expected_model, monkeypatch: pytest.MonkeyPatch):
     monkeypatch.setattr(
         "mutalyzer_retriever.parsers.json_ensembl._seq_from_rest",
